refactor: cart item bug
