+++ mobile/src/entities/index.ts
@@ -1,3 +1,4 @@
 export * from './category';
 export * from './product';
 export * from './user';
+export * from './cartItem';
