+++ mobile/src/shared/stores/OrderStore.ts
@@ -1,8 +1,6 @@
-import {create} from 'zustand';
 import axios from 'axios';
-
-import {CartItemType} from 'entities/CartItem';
-import {Product} from 'entities/product';
+import {CartItemType, Product} from 'entities';
+import {create} from 'zustand';
 
 export type OrderItem = {
   id: number;
@@ -18,13 +16,10 @@ export type Order = {
 };
 
 type OrderStore = {
-  isLoading: boolean;
-  isOrderItem: boolean;
   orders: Order[];
-
-
+  isLoading: boolean;
   setIsLoading: (isLoading: boolean) => void;
-  setIsOrderItem: (isOrredItem: boolean) => void;
+  isProductOrdered: (productId: number) => boolean;
   makeOrder: (
     items: CartItemType[],
     userid: number,
@@ -33,24 +28,25 @@ type OrderStore = {
   getOrders: (userid: number, token: string) => Promise<void>;
 };
 
-export const useOrderStore = create<OrderStore>(set => ({
+export const useOrderStore = create<OrderStore>((set, get) => ({
   isLoading: false,
-  isOrderItem: false,
   orders: [],
 
   setIsLoading: (isLoading: boolean) => {
     set({isLoading});
   },
 
-  setIsOrderItem: (isOrderItem: boolean)=> {
-    set({isOrderItem});
+  isProductOrdered: (productId: number) => {
+    const orders = get().orders;
+    return orders.some(order =>
+      order.orderItems.some(orderItem => orderItem.product.id === productId),
+    );
   },
 
   makeOrder: async (items: CartItemType[], userid: number, token: string) => {
     try {
       const formattedItems = items.map(item => ({
         quantity: item.quantity,
-
         productid: item.id,
       }));
       const {data} = await axios.post(
@@ -74,6 +70,7 @@ export const useOrderStore = create<OrderStore>(set => ({
       console.error(error.response);
     }
   },
+
   getOrders: async (userid: number, token: string) => {
     try {
       const {data} = await axios.get(
@@ -120,7 +117,6 @@ export const useOrderStore = create<OrderStore>(set => ({
       set({
         orders: formattedOrders,
       });
-    } catch (error: any) {
-      
-  }},
+    } catch (error: any) {}
+  },
 }));
